chore: rename config.json & defaults.json (review suggestions)
