+++ framework/cdv-gradle-config-defaults.json
